+++ miso-web/src/main/java/uk/ac/bbsrc/tgac/miso/webapp/controller/EditLibraryController.java
@@ -348,7 +348,7 @@ public class EditLibraryController {
       }
       dto.setParentSampleId(item.getId());
       dto.setParentSampleAlias(item.getAlias());
-      dto.setParentSampleProjectId(item.getProject().getId());
+      dto.setProjectId(item.getProject().getId());
       dto.setBox(newBox);
       return dto;
     }