+++ miso-web/src/main/webapp/scripts/form.js
@@ -312,6 +312,13 @@ FormUtils = (function($) {
         }
         return item;
       });
+    },
+
+    getSelectedTableData: function(containerId) {
+      if (initializedTables.indexOf(containerId) === -1) {
+        return null;
+      }
+      return ListState[containerId + 'Table'].selected;
     }
   };